feat(interface): better organisation in pages
