+++ app/Dockerfile
@@ -25,4 +25,4 @@ EXPOSE 8501
 
 HEALTHCHECK CMD curl --fail http://localhost:8501/_stcore/health
 
-ENTRYPOINT ["streamlit", "run", "Accueil.py", "--server.port=8501", "--server.address=0.0.0.0"] 
+ENTRYPOINT ["streamlit", "run", "🔍_Critères_d'éligibilité_par_AOM.py", "--server.port=8501", "--server.address=0.0.0.0"]
